refactor(schemas): migrate Phase 9 endpoint schemas — 6 files, 52 classes (#6042)
- agent.py: migrate GoalPayload, CommandApprovalPayload, EnhancedGoalPayload,
  MultiAgentTaskPayload, AgentAnalysisRequest, ResearchTaskRequest → schemas_agent.py
- agent_terminal.py: migrate 7 Terminal* request models → schemas_system.py
- chat_sessions.py: migrate SessionCreate, SessionUpdate, ActivityCreate,
  ActivityBatchCreate, SessionShareRequest, ChatResetRequest → schemas_chat.py
- code_search.py: migrate CodeSearchIndexRequest, CodeSearchRequest,
  CodeSearchGetResponse, CodeAnalyticsRequest, CodeDeclaration, ReusabilityReport,
  CodeSearchResponse → schemas_code.py
- entity_extraction.py: migrate ExtractionMessage, EntityExtractionRequest,
  BatchExtractionRequest, EntityExtractionResponse, BatchExtractionResponse,
  EntityExtractionHealthResponse → schemas_knowledge.py; v1→v2 validators
- git_mcp.py: migrate 6 Git* request models → schemas_code.py
