+++ autobot-backend/api/agent_terminal.py
@@ -229,7 +229,6 @@ import logging
 from typing import Dict, Optional
 
 from fastapi import APIRouter, Depends, HTTPException
-from pydantic import BaseModel, Field
 
 from auth_middleware import get_current_user
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
@@ -243,6 +242,14 @@ from api.schemas_agent import (
     AgentTerminalExecuteResponse,
     AgentTerminalInterruptResponse,
     AgentTerminalResumeResponse,
+)
+from api.schemas_system import (
+    TerminalApproveCommandRequest,
+    TerminalCreateSessionRequest,
+    TerminalExecuteCommandRequest,
+    TerminalHostSelectionRequest,
+    TerminalInterruptRequest,
+    TerminalToolApprovalRequest,
 )
 from api.schemas_terminal import (
     AgentTerminalCommandStateResponse,
@@ -263,103 +270,6 @@ logger = logging.getLogger(__name__)
 
 # Create router
 router = APIRouter(prefix="/agent-terminal", tags=["agent-terminal"])
-
-
-# Request/Response Models
-
-
-class CreateSessionRequest(BaseModel):
-    """Request to create agent terminal session"""
-
-    agent_id: str = Field(..., description="Unique identifier for the agent")
-    agent_role: str = Field(
-        ...,
-        description="Role of the agent (chat_agent, automation_agent, system_agent, admin_agent)",
-    )
-    conversation_id: Optional[str] = Field(
-        None, description="Chat conversation ID to link"
-    )
-    host: str = Field(
-        "main",
-        description="Target host (main, frontend, npu-worker, redis, ai-stack, browser)",
-    )
-    metadata: Optional[Dict] = Field(None, description="Additional session metadata")
-
-
-class ExecuteCommandRequest(BaseModel):
-    """Request to execute command in agent session"""
-
-    command: str = Field(..., description="Command to execute")
-    description: Optional[str] = Field(
-        None, description="Description of what command does"
-    )
-    force_approval: bool = Field(
-        False, description="Force user approval even for safe commands"
-    )
-
-
-class ApproveCommandRequest(BaseModel):
-    """Request to approve/deny pending command"""
-
-    approved: bool = Field(..., description="Whether command is approved")
-    user_id: Optional[str] = Field(None, description="User who made the decision")
-    comment: Optional[str] = Field(
-        None, description="Optional comment or reason for the decision"
-    )
-    auto_approve_future: bool = Field(
-        False, description="Auto-approve similar commands in the future"
-    )
-    # Permission v2: Project memory fields
-    remember_for_project: bool = Field(
-        False, description="Remember approval for this project"
-    )
-    project_path: Optional[str] = Field(
-        None, description="Project path for approval memory"
-    )
-
-
-class ToolApprovalRequest(BaseModel):
-    """Request to approve/deny a pending agent tool (event-stream level approval)."""
-
-    approved: bool = Field(..., description="Whether the tool execution is approved")
-    comment: Optional[str] = Field(None, description="Optional reason for the decision")
-    task_id: Optional[str] = Field(
-        None, description="Task ID from the APPROVAL_REQUIRED event"
-    )
-
-
-class InterruptRequest(BaseModel):
-    """Request to interrupt agent and take control"""
-
-    user_id: str = Field(..., description="User requesting control")
-
-
-class HostSelectionRequest(BaseModel):
-    """Request for agent to select an infrastructure host"""
-
-    agent_session_id: Optional[str] = Field(
-        None, description="Agent terminal session ID"
-    )
-    command: Optional[str] = Field(None, description="Command to execute on host")
-    purpose: Optional[str] = Field(None, description="Purpose of the SSH action")
-    preferred_host_id: Optional[str] = Field(
-        None, description="Preferred host ID if any"
-    )
-    allow_auto_select: bool = Field(
-        True, description="Allow auto-selection if default host is set"
-    )
-
-
-class HostSelectionResponse(BaseModel):
-    """Response to host selection request"""
-
-    request_id: str = Field(..., description="Unique request ID for tracking")
-    status: str = Field(..., description="pending_selection, selected, or cancelled")
-    selected_host_id: Optional[str] = Field(None, description="Selected host ID")
-    selected_host_name: Optional[str] = Field(None, description="Selected host name")
-    connection_info: Optional[Dict] = Field(
-        None, description="Connection details (host, port, username)"
-    )
 
 
 # Dependency for AgentTerminalService
@@ -413,7 +323,7 @@ def get_agent_terminal_service(
 @router.post("/sessions", response_model=AgentTerminalSessionCreateResponse)
 async def create_agent_terminal_session(
     current_user: dict = Depends(get_current_user),
-    request: CreateSessionRequest = None,
+    request: TerminalCreateSessionRequest = None,
     service: AgentTerminalService = Depends(get_agent_terminal_service),
 ):
     """
@@ -592,7 +502,7 @@ async def delete_agent_terminal_session(
 async def execute_agent_command(
     current_user: dict = Depends(get_current_user),
     session_id: str = None,
-    request: ExecuteCommandRequest = None,
+    request: TerminalExecuteCommandRequest = None,
     service: AgentTerminalService = Depends(get_agent_terminal_service),
 ):
     """
@@ -635,7 +545,7 @@ async def execute_agent_command(
 async def approve_agent_command(
     session_id: str,
     current_user: dict = Depends(get_current_user),
-    request: ApproveCommandRequest = None,
+    request: TerminalApproveCommandRequest = None,
     service: AgentTerminalService = Depends(get_agent_terminal_service),
 ):
     """
@@ -680,7 +590,7 @@ async def approve_agent_command(
 @router.post("/tools/approve/{approval_id}", response_model=AgentTerminalToolApprovalResponse)
 async def submit_tool_approval(
     approval_id: str,
-    request: ToolApprovalRequest,
+    request: TerminalToolApprovalRequest,
     current_user: dict = Depends(get_current_user),
 ):
     """Publish an APPROVAL_RESPONSE event for an agent tool approval request.
@@ -725,7 +635,7 @@ async def submit_tool_approval(
 async def interrupt_agent_session(
     session_id: str,
     current_user: dict = Depends(get_current_user),
-    request: InterruptRequest = None,
+    request: TerminalInterruptRequest = None,
     service: AgentTerminalService = Depends(get_agent_terminal_service),
 ):
     """
@@ -939,7 +849,7 @@ _pending_host_selections: Dict[str, Dict] = {}
 @router.post("/host-selection/request", response_model=AgentTerminalHostSelectionRequestResponse)
 async def request_host_selection(
     current_user: dict = Depends(get_current_user),
-    request: HostSelectionRequest = None,
+    request: TerminalHostSelectionRequest = None,
 ):
     """
     Agent requests host selection for SSH action.
