+++ autobot-backend/api/chat_sessions.py
@@ -8,8 +8,6 @@ from typing import Dict, List, Optional
 
 from fastapi import APIRouter, Depends, Request, Response
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel, Field
-
 from auth_middleware import get_auth_middleware, get_current_user
 from autobot_memory_graph import AutoBotMemoryGraph
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
@@ -21,16 +19,22 @@ from type_defs.common import Metadata
 from api.schemas_common import DataResponse
 from api.schemas_chat import (
     ActivityAddData,
+    ActivityBatchCreate,
     ActivityBatchData,
+    ActivityCreate,
     ChatResetData,
+    ChatResetRequest,
     SessionActivitiesData,
     SessionCheckpointClearData,
+    SessionCreate,
     SessionCreateData,
     SessionDeleteData,
     SessionListData,
     SessionMessagesData,
     SessionShareData,
     SessionSharePreviewData,
+    SessionShareRequest,
+    SessionUpdate,
     SessionUpdateData,
 )
 
@@ -137,72 +141,6 @@ def log_request_context(request, endpoint, request_id):
     """Log request context for debugging"""
     logger.info(
         "[%s] %s - %s %s", request_id, endpoint, request.method, request.url.path
-    )
-
-
-# ====================================================================
-# Request/Response Models
-# ====================================================================
-
-
-class SessionCreate(BaseModel):
-    """Session creation model"""
-
-    title: Optional[str] = Field(None, max_length=200, description="Session title")
-    team_id: Optional[str] = Field(
-        None, description="Team ID for team-scoped sessions (#684)"
-    )
-    metadata: Optional[Metadata] = Field(
-        default_factory=dict, description="Session metadata"
-    )
-
-
-class SessionUpdate(BaseModel):
-    """Session update model"""
-
-    title: Optional[str] = Field(None, max_length=200, description="New session title")
-    metadata: Optional[Metadata] = Field(None, description="Updated metadata")
-
-
-# Issue #608: Activity tracking models
-class ActivityCreate(BaseModel):
-    """Single activity creation model"""
-
-    activity_id: str = Field(..., description="Frontend-generated activity ID")
-    type: str = Field(
-        ..., description="Activity type: terminal, file, browser, desktop"
-    )
-    user_id: str = Field(..., description="User who performed the activity")
-    content: str = Field(
-        ..., max_length=10000, description="Activity content/description"
-    )
-    secrets_used: list[str] = Field(default_factory=list, description="Secret IDs used")
-    metadata: Optional[Metadata] = Field(
-        default_factory=dict, description="Activity metadata"
-    )
-    timestamp: str = Field(..., description="ISO format timestamp from frontend")
-
-
-class ActivityBatchCreate(BaseModel):
-    """Batch activity creation model"""
-
-    activities: list[ActivityCreate] = Field(
-        ..., description="List of activities to create"
-    )
-
-
-# Issue #689: Session sharing model
-class SessionShareRequest(BaseModel):
-    """Request to share a session with users"""
-
-    share_with: list[str] = Field(
-        ..., min_length=1, description="User IDs to share with"
-    )
-    include_knowledge: bool = Field(
-        False, description="Include KB facts from this session"
-    )
-    knowledge_facts: Optional[list[str]] = Field(
-        None, description="Specific fact IDs to share (all if omitted)"
     )
 
 
@@ -1494,16 +1432,6 @@ async def export_session(session_id: str, request: Request, format: str = "json"
 # =============================================================================
 # Issue #549: Chat Reset Endpoint
 # =============================================================================
-
-
-class ChatResetRequest(BaseModel):
-    """Request model for chat reset"""
-
-    session_id: Optional[str] = Field(
-        None, description="Session ID to reset (optional)"
-    )
-    clear_context: bool = Field(True, description="Clear conversation context")
-    keep_system_prompt: bool = Field(True, description="Keep system prompt after reset")
 
 
 def _preserve_system_messages(chat_manager, session_id: str) -> List[Dict]:
